Path for I/O is configurable now.
Before, only the puma-em directory could contain the temporary
directories that contained the data generated during the computations.
This has changed now: the default directory is defined in run.sh, but
can be passed as a variable to run.sh.

Another variable that is passed directly to run.sh is the number of
processes on which one desires to run the code.