Removed full mesh broadcast in mpi_mlfma.cpp
The mesh is read on process 0, then sliced into local meshes which are
then sent to the other processes.

Only the local meshes are used to create the arrays of the
octtree, and considerable memory savings are made, especially when
the problem becomes large, since the full mesh is not duplicated on each
process.

Significant related modifications all over the place as well.