Computation of needed RWG indexes for near field
We don't want to broadcast the full solution vector to all processes
anymore. So, we have to determine which indexes are needed locally for
Znear and SAI preconditioning matrices. This is done thanks to the new
code, RWGs_renumbering.cpp, which uses STL container multimap to find
the occurences of a particular RWG number and assign to all its occurences
a local RWG number.