Wrote compute_Z_near in C++
Not much of a speed enhancement, since most of the time is spent in
the Z calculation routine itself (there was not much Python overhead, in
other words). But it might allow a better synergy between C++
executables in the future, allowing for not passing through the disk for
I/O between programs, for example.