Enhanced mem management in mesh_functions_seb.cpp
A lot of memory was used by std::vector objects. Even after a clear() of
an object, the memory was still allocated to the object. Therefore, a
swap technique is used after clear, to reduce memory consumption. A more
radical solution will be welcome, such as a killing of the object pure
and simple.

Noticed a bug as well: the triangles vertexes were not reordered in
Python when calling mesh_functions_seb in mesh_functions_seb_C.py which
in turn is called by meshClass.py. It had no consequences, but it is now
corrected.